Readme (#18)
Adding support for README files.

A new leaf node type, FILE, gives packages a way to include unmodified
files. In this first implementation, build.py only looks for a single
file, with name README, that is expected to contain a description of
the package.

To support raw files, object files are now stored without a
format-specific extension (since raw files are format independent).