Disable the progress bar in the failover code (#27)
The "python" CSV engine is pretty broken, and working around its bugs is
probably not worth it. (E.g., it requires the file to be open in text mode,
not binary, even though that's not documented. But text mode will almost
certainly break .xls and other binary files.)

Also add a test CSV with invalid unicode that triggers the failover.