Adding RootNode to core.py and package schema
Also removing the cli container from the docker test environment. That
container is huge and takes both a long time to build and a lot of
space in the docker VM. Now that we can map all the ports to localhost,
we don’t need it to run the integration test.