Command args to generate build file without building the package (#85)
#### Command args to generate build file without building the package

This addition supports a common pattern: generating a build file for a
collection of sources then editing that build file (e.g. to specify
package format or per-source transforms) before building a package.

Making generate (a build-file from source) a separate command instead
of making it an option for quilt build. Generate doesn’t actually build
a package so it doesn’t need a package name.

Also renaming the directory (-d) option in quilt build to auto (-a).