Use file extension to avoid duplicates (#87)
#### Adding file extensions to node names to resolve duplicate names

In cases where nodes would otherwise end up with duplicate names in auto-generated buildfiles, the buildfile generator now includes files' extensions to avoid name collisions. Some collisions are still possible (and detected), but most be will be resolved assuming that source files are uniquely named.

Also changing the arguments to quilt build so that instead of requiring an explicit argument to generate a build file, build now looks to see if path is a directory and then if that directory contains build.yml. Build.yml will be generated if it's not already present.