Add support for accessing groups as data
Particularly for spark, but also generally for partitioned data, we
want to support accessing datasets at different granularities. The code
here allows us to combine a set of nodes (only 1-level down in the
package tree) into a common table. To support general (multi-level)
partitioning, we’ll need to extend this to grab all data nodes below
the queried node and validate that the schemas match.