Share packages by email (#103)
Allows sharing packages via email. A sharing request (access_put) with an email address instead of a username creates an invitation object that is registered in the database. Flask uses Django to send the invitation email. Invitations are converted to package access when users view their profile. Typically an email address is used when the package recipient doesn't yet have a Quilt account. They receive access to the packages to which they've received invitations when they first login after creating their account.