Refactor package build interface (#175)
This changes the way Quilt packages are stored locally. Instead of looking for quilt_packages in the current working directory, Quilt now uses a single package repository that can hold multiple instances of a package.

The new directory structure contains both the raw data and metadata for packages including any tags and versions associated with each package instance. Using the file system to store this metadata is expected to be slower than a database or single-file lookup, but it is thread-safe and human-readable and requires no additional dependencies.

Copying the behavior of docker images, quilt ls displays a separate
line for each tag when a given instance has multiple tags.