Optionally use subprocess in CLI API testing
There are two methods of testing the CLI API -- directly calling
tools.main.main(), or by calling quilt in a subprocess.  The former
is significantly faster due to current quilt load times (probably
pandas and pyarrow).  Calling via a subprocess could be argued to
be more complete, so I've used an OS env var as a switch.  When
QUILT_TEST_CLI_SUBPROC exists and is "true" (any case), then testing
will use a subprocess instead of the default, faster method.