Build cache (#187)
Basic implementation of a build-cache for TableNodes in packages. The build cache stores a mapping from source files (with associated processing instructions, e.g., transform, etc.) to the resulting binary fragments that are included in the final package. This saves the most time-consuming step of quilt build, converting text inputs and serializing to Parquet.