Keep build_from_cache in sync with test/build.yml
The test test_build_from_cache contains an assertion that verifies that
the cache is actually filled. The hash key in the cache is generated
from both the source file and kwargs so it needs to be kept in sync
with the build file (test/build.yml).

Fix an indentation bug in test/build.yml.