Use a thread-local session when making requests to the auth server
Right now, we're creating a new connection for every auth request, which is pretty inefficient (TCP handshake, TLS, etc.). Sessions allow us to use a connection pool - however, they're not quite thread safe. So, use a session per thread.

(Tested it locally using uwsgi with the production config - and it appears to be doing the right thing.)