Don't expire objects on commit
Turns out, by default, SQLAlchemy expires all objects on commit, and causes them to be re-queried if they're accessed again.
It's particularly bad in case of `package_get`: we commit an audit event, then access blob sizes. If we have 10,000 blobs, SQLAlchemy will issue 10,000 individual queries to look up the size of each one.

We don't depend on the objects getting updated when they're commited, to let's disable this "feature".