Compose packages (#388)
Adds a new build-file keyword, "package", that includes a whole package or sub-package tree in the package being built. This allows composing packages out of other packages and fragments of other packages. It also supports using small single-node packages as a form of object-store from which Quilt packages can be built.