Upgrade SQLAlchemy and use the new "pre-ping" option (#509)
Currently, if the DB connection goes away (DB gets restarted, network settings change, etc.), SQLAlchemy will still try to use the stale connection, fail, and remove the stale connection out of the pool. However, because we have 16 registry processes, we can expect 16 exceptions (each one causing a 500 error) when this happens.
SQLAlchemy has a new option to "pre-ping" a connection (using a `SELECT 1` query) before trying to use it. It seems to have negligible overhead, so let's use it.

Of course, upgrading SQLAlchemy is a slightly risky change.