Allow asa calls on non-table GroupNodes (#626)
* Allow asa calls on non-table GroupNodes

Previously, calling _data on GroupNodes was only allowed if all the
node’s children were TableNodes with matching schemas. This change
allows _data to be called with an asa function to return all child-node
paths regardless of their type. Updated docs to describe passing an asa function on a GroupNode that's not a DataFrame.