Don't skip header row when reading csv/excel files in unit tests
Skipping the header in Pandas read_csv and read_excel can produce DataFrames with funky column names and indexes, the testing of which is outside the scope of the current unit tests that specify the line skips.

In future, we should add Parquet/Pandas round-trip tests that cover all types of Pandas DataFrames, including uncommon indexing and column names.