Remove package class (#731)
This change set simplifies the codebase by removing the Package class (and package.py). Package was one of three different representations of a package in the code. After this change, packages are represented as trees of nodes--either from nodes.py or core.py. Unifying those sets of classes is still TBD. Most of the functionality from the Package class (package.py) was moved to the PackageStore (store.py).