Implement an S3 Select signing proxy
It just signs the request and forwards to S3.

For now, the bucket region is set using an env var. It's probably good enough for the analytics bucket.
(We _could_ look it up, but that's an extra rount trip to S3. Maybe the catalog could send it as a parameter.)

Some fixes to the shared code:
- Support POST and OPTIONS
- Support for proxy resources ("/s3select/bucket/path/..." instead of just "/s3select")
- Fix a bug in CORS