Implement an S3 Select signing proxy (#1015)
It just signs the request and forwards to S3.

For now, the bucket region is set using an env var. It's probably good enough for the analytics bucket.
(We _could_ look it up, but that's an extra rount trip to S3. Maybe the catalog could send it as a parameter.)