Fix listing objects in the s3 proxy (#1352)
AWS SDK does not include a `/` after the bucket name when listing objects.
Fix the regex to not require `/`, and also stop at `?` so we don't drop the query string.

S3 server does not seem to care if the `/` is present or not. The following produce the same result:
- `https://s3.amazonaws.com/quilt-example?list-type=2`
- `https://s3.amazonaws.com/quilt-example/?list-type=2`

But of course we can't just proxy everything starting with the `/` because these are NOT the same:
- `https://s3.amazonaws.com/quilt-example/dima/q/foo%3F.txt`
- `https://s3.amazonaws.com/quilt-example//dima/q/foo%3F.txt`

Sigh.