Use ChunksizeAdjuster to adjust chunk size (#1388)
...because of course it's a thing.

S3 allows a maximum of 10,000 chunks - and with the default chunk size of 8MB,
we would go over that if the file is bigger than 80GB.

Boto adjusts chunk sizes to deal with that, so let's do the same.

Also adjust unit tests to use realistic file sizes (because there's also
a minimum allowed chunk size). Test the >10,000 chunks case in multipart copy.