Fix exception handling when copying files
Currently, the code creates a list of futures, then waits for them in _reverse_ order. That's fine as long as they all succeed.

But if they all fail (access denied, out of disk, etc.), we will keep trying every single object, and only the last exception will propagate to the main thread.