es/indexer: Drop fatal 40X failures on the floor (#1804)
We choose to skip the dead letter queue in select cases where we are unlikely to recover from the error, so as to avoid failing the batch. We found a case where list_object_versions may return an object that will never be readable (e.g. due to a foreign owner).