Remove call to ES & optimize listing
Remove the calls to ES because they're no longer needed and won't work for buckets that aren't indexed.

Use a single pass through the pointers directory for speed.

TODO:
It's still painfully slow for directories with many package revisions because we're reading the contents of all pointer files and doing it sequentially.