Add a region config variable to support GovCloud (#3685)
* Add a defaultRegion config variable to support GovCloud

GovCloud doesn't have us-east-1, so we need to explicitly set a valid region for S3. It's also a potential optimization for any non-us-east-1 stacks (assuming most S3 buckets are in the same region as the stack).

Also use it for Athena: seems less hacky that extracting the region from apiGatewayEndpoint.

* Make defaultRegion required

* try to fix tests

* Rename defaultRegion to region