Rebase the catalog Docker image from "nginx" to "amazonlinux"
This fixes vulnerabilities found by the ECR scanner.

The nginx base image contains a production-ready nginx.conf that sets a few sane defaults and loads everything else from the "conf.d" dir.
The default nginx.conf, on the other hand, configures a server on port 80 that displays "Welcome to nginx".
Therefore, we're now overwriting nginx.conf, too, in addition to adding the catalog config.

"server_tokens off;" is moved to the main config where it actually belongs.

A few things are added to the Dockerfile that the nginx base image Dockerfile used to do.