Move es/indexer/ to indexer/
It is the only lambda in the es/ dir, as well as the only lambda that is two directories deep.
Let's get rid of the "es" directory to keep everything consistent.

It also helps with deployment: in S3, we just use "indexer", not "es/indexer", so we need to call `basedir`.